chore(mobile): Simplified SendOutputsScreen and moved logic to useSendForm
